requirements.txt
# requirements.txt
# Base ------------------------------------------------------------------------
gitpython==3.1.43
matplotlib==3.5.3
numpy==1.23.5
opencv-python==4.9.0.80
Pillow==10.3.0
psutil==5.9.8
PyYAML==5.3.1
requests==2.31.0
scipy==1.4.1
torchvision==0.16.0
torchaudio==2.1.0
torch==2.1.0
efficientnet-pytorch==0.7.1
tqdm==4.66.2
ultralytics==8.1.42
scikit-learn==1.0.2
elasticsearch==8.12.1
kaggle==1.6.12
tensorflow==2.13.1

# Backend ------------------------------------------------------------------------
Django==4.2
drf-spectacular==0.27.2
djangorestframework==3.15.2

# Plot ------------------------------------------------------------------------
pandas==1.3.5
seaborn==0.13.2
comet_ml==3.39.3

# Deploy ----------------------------------------------------------------------
setuptools==68.2.2

# Extras ----------------------------------------------------------------------
wheel==0.41.2
ipython==8.12.3